move preprocessing functions to helper page and add sidebar info
